Add files via upload
This is guessing game, a small program implementing the basic HTML, CSS and basic javascript code. The user can play this game to find a unknown number between 1 and 100, based on the given hint whether the number is larger or smaller compared to user number and progress accordingly to find the actual number.